+++ src/web/components/QuickValuesPlusWidgetEditConfiguration.jsx
@@ -76,23 +76,14 @@ const QuickValuesPlusWidgetEditConfiguration = React.createClass({
                onChange={this.props.onChange}
                help="Include a table with quantitative information."/>
 
-        <Input key="displayAddToSearchButton"
-               type="checkbox"
-               id="quickvaluesplus-display-add-to-search-button"
-               name="display_add_to_search_button"
-               label="Display Add to Search Button"
-               checked={this.props.config.display_add_to_search_button}
-               onChange={this.props.onChange}
-               help="Column containing a add to search button." />
-
-        <Input key="displayRemoveFromSearchButton"
-               type="checkbox"
-               id="quickvaluesplus-display-remove-from-search-button"
-               name="display_remove_from_search_button"
-               label="Display Remove From Search Button"
-               checked={this.props.config.display_remove_from_search_button}
-               onChange={this.props.onChange}
-               help="Column containing a remove from search button." />
+          <Input key="displayTermHyperlinks"
+                 type="checkbox"
+                 id="quickvaluesplus-display-term_hyperlinks"
+                 name="display_term_hyperlinks"
+                 label="Display Term Hyperlinks"
+                 checked={this.props.config.display_term_hyperlinks}
+                 onChange={this.props.onChange}
+                 help="Turn off hyperlinks for the term." />
 
         <Input key="displayExludeFromQueryButton"
                type="checkbox"